+++ packages/components/src/components/notification-banner/notification-banner.tsx
@@ -9,9 +9,19 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-import { Component, h, Host, Prop, Element, Method } from '@stencil/core';
+import {
+  Component,
+  h,
+  Host,
+  Prop,
+  Element,
+  Method,
+  Event,
+  EventEmitter,
+} from '@stencil/core';
 import classNames from 'classnames';
 import statusNote from '../../utils/status-note';
+import { emitEvent } from '../../utils/utils';
 
 @Component({
   tag: 'scale-notification-banner',
@@ -28,6 +38,8 @@ export class NotificationBanner {
   @Prop() autoHide?: boolean = false;
   @Prop() autoHideDuration?: number = 3000;
   @Prop() href: string;
+  /** Fires when the notification banner has been dismissed */
+  @Event({ eventName: 'scale-close' }) scaleClose: EventEmitter<void>;
 
   hasSlotText?: boolean;
   hasSlotLink?: boolean;
@@ -63,6 +75,7 @@ export class NotificationBanner {
               class="notification-banner__icon-success"
               color="#187431"
               accessibility-title="success"
+              aria-hidden="true"
             />
           );
         case 'informational':
@@ -70,6 +83,7 @@ export class NotificationBanner {
             <scale-icon-alert-information
               class="notification-banner__icon-information"
               accessibility-title="information"
+              aria-hidden="true"
             />
           );
         case 'error':
@@ -77,6 +91,7 @@ export class NotificationBanner {
             <scale-icon-alert-error
               class="notification-banner__icon-error"
               accessibility-title="error"
+              aria-hidden="true"
             />
           );
         case 'warning':
@@ -84,7 +99,7 @@ export class NotificationBanner {
             <scale-icon-alert-error
               class="notification-banner__icon-information"
               color="#AE461C"
-              accessibility-title="information"
+              aria-hidden="true"
             />
           );
       }
@@ -94,6 +109,7 @@ export class NotificationBanner {
 
   close = () => {
     this.opened = false;
+    emitEvent(this, 'scaleClose');
   };
 
   render() {
@@ -104,6 +120,8 @@ export class NotificationBanner {
     return (
       <Host>
         <div
+          role="alert"
+          style={{ display: `${this.opened ? '' : 'none'}` }}
           part={this.getBasePartMap()}
           class={this.getCssClassMap()}
           tabindex="0"
@@ -115,15 +133,18 @@ export class NotificationBanner {
               {this.dismissible && (
                 <button
                   part="button-dismissable"
+                  type="button"
                   class="notification-banner__button-close"
                   onClick={() => this.close()}
+                  tabindex={0}
+                  aria-label="close"
                   onKeyDown={(e) => {
                     if (e.key === 'Enter') {
                       this.close();
                     }
                   }}
                 >
-                  <scale-icon-action-circle-close accessibility-title="close" />
+                  <scale-icon-action-circle-close />
                 </button>
               )}
               {this.hasSlotText && (
@@ -133,7 +154,11 @@ export class NotificationBanner {
               )}
 
               {this.hasSlotLink && (
-                <scale-link href={this.href} class="notification-banner__link">
+                <scale-link
+                  href={this.href}
+                  class="notification-banner__link"
+                  role="link"
+                >
                   <slot name="link" />
                 </scale-link>
               )}