+++ CHANGELOG.md
@@ -3,6 +3,59 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.104](https://github.com/telekom/scale/compare/v3.0.0-beta.103...v3.0.0-beta.104) (2022-05-12)
+
+
+### Bug Fixes
+
+* **date-picker:** keyboard nav styles ([#1034](https://github.com/telekom/scale/issues/1034)) ([6f30cd5](https://github.com/telekom/scale/commit/6f30cd54201c8043dccfb971a87dabc0f8c7c0fd))
+* css transition values ([#1033](https://github.com/telekom/scale/issues/1033)) ([eb71637](https://github.com/telekom/scale/commit/eb71637ee607d8be4bea038549adcfdbc4c2f327))
+
+
+### Features
+
+* **notification-toast:** add css var for z-index ([#1041](https://github.com/telekom/scale/issues/1041)) ([e4c9820](https://github.com/telekom/scale/commit/e4c982093a7f0339aac21d982ad6f64248cf38a7))
+
+
+
+
+
+# [3.0.0-beta.103](https://github.com/telekom/scale/compare/v3.0.0-beta.102...v3.0.0-beta.103) (2022-05-04)
+
+
+### Bug Fixes
+
+* **modal:** clicking esc key to close all opened modal window ([#1003](https://github.com/telekom/scale/issues/1003)) ([dbe17fd](https://github.com/telekom/scale/commit/dbe17fd46ae7d136be358a5699d878a18c7f939b))
+* **text-input:** remove padding, add font-weight to helper text ([#1019](https://github.com/telekom/scale/issues/1019)) ([373ca47](https://github.com/telekom/scale/commit/373ca4789e98f39ff593364df1a69314678d6805))
+* transition ([#1028](https://github.com/telekom/scale/issues/1028)) ([a36e97f](https://github.com/telekom/scale/commit/a36e97fe635472bfa9bc1cf11743f6f3e84f571a))
+* **header:** use valid li > a html nesting + a11y improvements ([#1015](https://github.com/telekom/scale/issues/1015)) ([614f83d](https://github.com/telekom/scale/commit/614f83d485bdd9fed05282b9ff5201bf9cbed804))
+* increase flyout zindex ([#1022](https://github.com/telekom/scale/issues/1022)) ([6e75493](https://github.com/telekom/scale/commit/6e75493625484319bd863323989431862596f18b))
+* remove transform to maintain fixed positioning in child elements ([#1006](https://github.com/telekom/scale/issues/1006)) ([2fea224](https://github.com/telekom/scale/commit/2fea224a3594aa60bbddd6383c5337c75d22e28d))
+* **storybook:** broken path in import snippet ([#1014](https://github.com/telekom/scale/issues/1014)) ([8724bfc](https://github.com/telekom/scale/commit/8724bfc669e5cad4bb8234148240c9727ff84e5f))
+
+
+### Features
+
+* **toggle-button:** add setFocus method ([#1005](https://github.com/telekom/scale/issues/1005)) ([8754b66](https://github.com/telekom/scale/commit/8754b663765f035e1344cadb6bdb4f012bed1435))
+
+
+
+
+
+# [3.0.0-beta.102](https://github.com/telekom/scale/compare/v3.0.0-beta.101...v3.0.0-beta.102) (2022-04-19)
+
+
+### Features
+
+* **button:** add name and value props ([#996](https://github.com/telekom/scale/issues/996)) ([32270da](https://github.com/telekom/scale/commit/32270dae2bf1dd6d463ff0bef0dcf14e033b2a32))
+* **modal:** add closeButtonTitle prop, closes [#771](https://github.com/telekom/scale/issues/771) ([#995](https://github.com/telekom/scale/issues/995)) ([1555fe3](https://github.com/telekom/scale/commit/1555fe3a25b2bcf3132ae664065de03de0c91cac))
+* **notification-toast:** add auto-hide prop, closes [#948](https://github.com/telekom/scale/issues/948) ([#979](https://github.com/telekom/scale/issues/979)) ([aa1c3be](https://github.com/telekom/scale/commit/aa1c3bef2c8a3267155f0a40eaed12a01bdb81e2))
+* **text-field:** add min and max attributes ([#988](https://github.com/telekom/scale/issues/988)) ([a073dfb](https://github.com/telekom/scale/commit/a073dfb3120997bc16cd9923cb2ec3b32d88aa37))
+
+
+
+
+
 # [3.0.0-beta.101](https://github.com/telekom/scale/compare/v3.0.0-beta.100...v3.0.0-beta.101) (2022-04-07)